+++ core/src/core/printfun.c
@@ -1,6 +1,0 @@
-#include "printfun.h"
-#include <stdio.h>
-
-void print_func(i32 num) {
-    printf("The number is: %d\n", num);
-}