+++ testbed/src/testbed.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <core/defines.h>
+#include <fracture/core/defines.h>
 #include <fracture.h>
 
 typedef struct testbed_state {